file_export: modularized WAV/AVI into container interface
* moved multimedia functions back into file_export; removed multimedia.c
* added MULTIMEDIA preprocessor symbol
* moved screenshot convenience function into file_export
* removed unneeded fwritele now that audio_pcm converts samples itself
