+++ src/codecs/container_avi.h
@@ -1,0 +1,9 @@
+#ifndef CODECS_CONTAINER_AVI_H_
+#define CODECS_CONTAINER_AVI_H_
+
+#include "atari.h"
+#include "codecs/container.h"
+
+extern CONTAINER_t Container_AVI;
+
+#endif /* CODECS_CONTAINER_AVI_H_ */
