+++ src/codecs/container_wav.h
@@ -1,0 +1,9 @@
+#ifndef CODECS_CONTAINER_WAV_H_
+#define CODECS_CONTAINER_WAV_H_
+
+#include "atari.h"
+#include "codecs/container.h"
+
+extern CONTAINER_t Container_WAV;
+
+#endif /* CODECS_CONTAINER_WAV_H_ */
