+++ src/ui.c
@@ -83,7 +83,7 @@
 #include "sound.h"
 #endif /* SOUND */
 #if defined(SOUND) || defined(AVI_VIDEO_RECORDING)
-#include "multimedia.h"
+#include "file_export.h"
 #endif /* defined(SOUND) || defined(AVI_VIDEO_RECORDING) */
 #ifdef DIRECTX
 #include "win32\main.h"
@@ -1278,11 +1278,11 @@ static void SoundRecording(void)
 		UI_driver->fMessage("Can't record. Sound not enabled.", 1);
 		return;
 	}
-	if (!Multimedia_IsFileOpen()) {
+	if (!File_Export_IsRecording()) {
 		char buffer[FILENAME_MAX];
-		if (Multimedia_GetNextSoundFile(buffer, sizeof(buffer))) {
+		if (File_Export_GetNextSoundFile(buffer, sizeof(buffer))) {
 			/* file does not exist - we can create it */
-			FilenameMessage(Multimedia_OpenSoundFile(buffer)
+			FilenameMessage(File_Export_StartRecording(buffer)
 				? "Recording sound to file \"%s\""
 				: "Can't write to file \"%s\"", buffer);
 			return;
@@ -1290,7 +1290,7 @@ static void SoundRecording(void)
 		UI_driver->fMessage("All sound files exist!", 1);
 	}
 	else {
-		Multimedia_CloseFile();
+		File_Export_StopRecording();
 		UI_driver->fMessage("Recording stopped", 1);
 	}
 }
@@ -1299,11 +1299,11 @@ static void SoundRecording(void)
 #ifdef AVI_VIDEO_RECORDING
 static void VideoRecording(void)
 {
-	if (!Multimedia_IsFileOpen()) {
+	if (!File_Export_IsRecording()) {
 		char buffer[FILENAME_MAX];
-		if (Multimedia_GetNextVideoFile(buffer, sizeof(buffer))) {
+		if (File_Export_GetNextVideoFile(buffer, sizeof(buffer))) {
 			/* file does not exist - we can create it */
-			FilenameMessage(Multimedia_OpenVideoFile(buffer)
+			FilenameMessage(File_Export_StartRecording(buffer)
 				? "Recording video to file \"%s\""
 				: "Can't write to file \"%s\"", buffer);
 			return;
@@ -1311,7 +1311,7 @@ static void VideoRecording(void)
 		UI_driver->fMessage("All video files exist!", 1);
 	}
 	else {
-		Multimedia_CloseFile();
+		File_Export_StopRecording();
 		UI_driver->fMessage("Recording stopped", 1);
 	}
 }
